Add SvelteKit skeleton
I ran `npm init svelte` in the frontend directory to create all of the
files necessary for a SvelteKit app. I made a few changes to some of the
defaults in the Prettier config to format the frontend code consistently
with our Python code (assuming we'll be using Black for Python). I also
merged the .gitignore contents into our root .gitignore.

Although SvelteKit is not out of beta, the remaining issues in their 1.0
milestone don't seem like they'll pose a problem for us - the only
breaking change listed is about AMP, which we don't use. I've also been
assured by several people who have been using SvelteKit for a while that
it's fairly stable at this point.

Since SvelteKit is The Way of the Future(TM) it seems prudent to set it
up now in anticipation of a 1.0 later this year, rather than trying to
convert the Vite Svelte SPA template into a SvelteKit app later. Using
SvelteKit also leaves open the possibility of some server-side rendering
using a Lambda adapter in the future.