Add USWDS
I added USWDS, Sass, and svelte-preprocess to the project with a basic
stylesheet that pulls in all of USWDS (for now). We'll probably want to
try pairing down the things we import, once this is working.

I can't figure out how to configure additional include paths for Sass.
There's nothing I can find in the svelte-preprocess docs that mention
it, and my guess based on how I've done it in the past with Svelte,
Rollup, and svelte-preprocess-sass have not worked.