Configure SCSS for Vite and Svelte
I created an `scss` object to contain the configuration for Sass because
it needs to be set for both Vite and SvelteKit. In addition to the
Svelte preprocessor, I added the config to the Vite config so we can
import the styles in a script tag in Svelte components. This way we get
a stylesheet that isn't run through Svelte's aggressive unused selector
pruning and subject to Svelte's scoped styles. We could add PostCSS and
use svelte-preprocess's global attribute, but my feeling is an approach
that has fewer dependencies is preferable.