Configure static adapter for prerender
We're going to prerender the whole app (for now) and treat it as an SPA.
We should be able to rely on people using this having JavaScript
enabled and SEO is not really a concern, so an SPA should be ok, at
least for now. We may be able to eek some performance benefits out of
some server-side rendering, but we'll cross that bridge when we come to
it.