Configure alias for loading fonts in development
I added an alias for `/fonts` to the Vite configuration so that it can
find the font files in `node_modules/` during development. I'm not able
to set up a copy of the fonts as part of the build, so I think I'll have
to copy these in the Docker container or CI, depending on how we decide
to host the application.