Create a base layout for the application
I added a base layout in src/routes so that we can set the header and
footer for the app in a common location (in case we end up developing
multiple pages). I replaced the <h1> from the skeleton app with a
header component.

I had to add an exception for src/lib in the root .gitignore because
SvelteKit has a built-in alias so that it's easy to get to components
and helpers by importing from '$lib', but lib is ignored in the default
Python .gitignore.