Create a skeleton flask app
The app has a single route that just returns a hard-coded JSON object
for testing. This will enable us to determine whether we have the
service running or not. It has a rough structure that the Flask app will
use—including logging, the [app factory pattern][1], and a WSGI module
for use with Gunicorn.

The routes module may get refactored, depending on how complex the app
becomes. We may adopt a more Django-like structure with purpose-specific
modules (e.g. diag/routes.py, forecasts/routes.py, etc.).

[1]: https://flask.palletsprojects.com/en/2.1.x/patterns/appfactories/