Configure a Docker image to run Gunicorn
Add a Dockerfile to build an image with Gunicorn and our
unified_graphics app installed. By default it serves on port 8000,
although perhaps this should be port 80 to mirror the image configured
for the UI. Since this will all be running in a cluster on a private
subnet, that's probably fine.