Add a pre-commit hook for linting/formatting
I'd really like to avoid commits whose sole purpose is to fix formatting
issues, so I've written a pre-commit bash script that checks all of the
files staged for commit for linting and formatting errors to prevent
them being committed before they are fixed.

This will be optional for other developers, but it can be easily
configured by running

    git config core.hookspath .githooks

I'll include instructions for doing so in the CONTRIBUTING.md file.