Set default indentation to two spaces
I changed .editorconfig so that all files default to two space
indentation except for Python files. The pre-commit script uses two
spaces (per Google's shell style guide) and I was fighting with VSCodium
the whole time because editorconfig was configured to use four spaces.
Since shell scripts will often not have file extensions, there wasn't an
easy way to add them to the block of two-space files.

I had to reformat some of the files in the UI service as well, which is
for the best, I think.