Add typescript to linting
We don't have much typescript, but there is one file for typescript
definitions, so we might as well lint those. I also added a note about
some of the files that are missed by our pre-commit script. This may not
be a big problem, since these files will likely not change often, so
even if bad formatting does slip through, there won't be many
"Formatting" commits.