Setup GitHub Actions for the API service
Add two workflows:
1. To lint, type-check, test, build and scan the API service
2. To clean up PR images for closed PRs

The first workflow will build PR images for preview deployments as well
as images for the main branch and for tags like `x.y.z` or `x.y.z-rc#`.
It will also use Trivy to scan for critical & high vulnerabilities in
the container image and upload them for viewing in GitHub's Security
tab.