Get the diag_temperature test passing
Return a hard-coded response to match what the test expects. I need to
figure out a way to provide known test data for this test so I can
assert the contents of the entire response, including the values,
rather than just checking to make sure the structure is correct
(property keys and types) without testing the actual values.