Add a diag module for retrieving data from NetCDF
In order to help us test the API routes with known data, I'm adding a
layer of abstraction between the route itself and the data loading.
We'll use this diag.get_diagnostics function to retrieve the
xarray.Dataset from the filesystem (or S3 or whatever), this way we can
set up a contract about what parameters are needed to find the data and
how it's formatted on return without tying ourselves to the actual
NetCDF files on disk. This way, if things change, we don't have to
rewrite a bunch of route functions.

It also means we can just mock this module to supply test data for
testing the routes so we don't need to commit any test NetCDF files to
run unit tests.