Return datasets from get_diagnostics
get_diagnostics needs to return the Dataset for both the guess (first
minimization loop) and the analysis (third minimization loop) of the
model initialization. I've added a test to ensure that we get a tuple of
Datasets back and added some type annotations so that you can tell—at
least a little—what get_diagnostics() is returning.