Mock get_diagnostics in test_temperature_diag_distribution
Mock the get_diagnostics function in test_temperature_diag_distribution
to ensure that it's been called. Ultimately I don't think we'll leave
this assertion in place, since it's not what we really care about and
we'll be able to assert the API response to ensure things are working,
but for right now this gets the test failing so that we have to update
the route function.