Generate histogram from dataset in diag_temperature
Replaced the hard-coded JSON response in the diagnostics route with code
that uses the return value from get_diagnostics to generate the
histogram, observation count, standard deviation, and mean from the
data.

This could probably use a refactor, as I think it's likely that we want
the diagnostics module to handle these calculations and just let the
routes return these values.