Rewrite diag tests to expect processed diagnostics
I think it makes more sense to have the diag module handle processing
the diagnostics files into dictionaries the API can use, so I've changed
the asserts to expect a dictionary instead of the Dataset.

I've also switched to using pytest's monkeypatch fixture instead of
unittest.mock. When in Rome.