Build the diag object we care in get_diagnostics
Instead of returning two Datasets from get_diagnostics, return a
dictionary that contains the statistics about the observations that we
care about.

This has a revealed a major flaw in how I mocked the data for the API
routes, because those tests are still passing, despite the fact that I
have changed the return type of get_diagnostics.