Mock open_dataset, not get_diagnostics
We're only mocking functions to avoid having to provide NetCDF files as
part of our test fixtures, so we should mock as close to that level as
possible. By mocking get_diagnostics, we decoupled the route from a
function it depends on so that when that function changed our test still
passed. By mocking open_dataset, we can still provide test data without
NetCDF files on disk but if a change happens to get_diagnostics, the
route tests will fail.