Add a test for a new get_filepath function
To save myself the trouble of having a complex mock for open_dataset
that returns different Dataset objects based on the filename passed to
the mocked function, I'm going to add a function that generates the
paths to diag files which can be used from get_diagnostics. There is
some risk here that a change to get_filepath that breaks get_diagnostics
will be missed, but that is arguably the purview of an integration test.