Add vitest and testing-library/svelte for unit testing
Separate out integration tests and unit tests for components. Vitest and
@testing-library/svelte were the recommendations I found on the Svelte
discord for this. I tried Playwright's experimental component testing
feature, but I couldn't get it going in this project, so we'll limit our
use of Playwright to integration tests right now. Maybe we can revisit
the component testing in Playwright later.