Add diag statistics to the Observations component
Display observation count, mean, and std. dev. in the Observations
component.

I have tests for this, but I don't like them. They guarantee the text is
in the DOM, but they don't guarantee that it's visible, or that it's
paired with the right label. This may be a good use for a test id. For
now, I'm going to leave it as at least guaranteeing that the data gets
into the DOM.