Show the histogram in the variable diag component
Move the Histogram component into VariableDiag. For now it's only being
uesd here, I don't know if it will be needed anywhere else (probalby it
will), but I'd rather refactor my charts later when I have an idea of
what their common components will be instead of guessing.

I don't have tests for the histogram at the moment, because I'm at a
loss for how to test it.