Use __DIAG_API_HOST__ to fetch data
I refactored the fetch call from routes/index.svelte into a helper
module for the api. The helper module defines a constant,
`DIAG_API_HOST` which is set to the string `__DIAG_API_HOST__`, which
gets replaced by vite. This allows us to dynamically configure the
location of the API when we build the application.