Remove vite-plugin-replace
Ian prefers providing the API host at run time, not at build time
because it makes the container more portable—i.e. we don’t have to build
a separate container for each environment. He's right about that.

Additionally, connecting from the client without a proxy means the
Python application has to be public, which is an extra step. So instead
we're going to switch to using a SvelteKit adapter that can run on the
server and handle API requests from the client.