Configure the SvelteKit Node adapter
I replaced the static adapter with the Node adapter so we get a Node
server that hosts the frontend and will have routes that we can use to
proxy requests to the Python API. This opens up some possibilities for
pre-rendering pages and possibly moving some work out of Python and into
SvelteKit—i.e. the Python application could become just a thin client
for dealing with NetCDF/Zarr files, if that is valuable.