Pass bins as test data in the VariableDiag tests
Now that we're building a shared domain and range for the charts, we
need to have bins available in the VariableDiag component when we run
tests or we get an exception when we try to call bins.concat(). This is
probably a sign that the tests are bad because we aren't passing
realistic data.