Install production dependencies in Docker image
After we copy the built SvelteKit app to the clean Node image, we need
to reinstall non-dev dependencies. This wasn't a problem before because
d3 is used client-side only, so it gets bundled into the client scripts.
But debug needs to be in the server environment because it's used by the
routes.

I had to pass --ignore-scripts to npm install so that npm wouldn't try
to run npm prepare, which requires the svelte-kit CLI to be installed,
which is only a dev dependency. I'm not sure this is the best way to
handle this...