Implement wind diag API as a wrapper around diag.wind
I created a dataclass to represent diagnostic data for vector data so
that we can benefit from more type checking. This means the
responsibility for converting the dataclass into JSON is the
responsibility of the API. Right now the diag.wind function doesn't do
anything other than return an empty VectorDiag object.