Change the response schema for wind diag data
I think it makes more sense to send wind data as direction and magnitude
rather than u, v vectors, since most of what we'll be visualizaing is
direction and magnitude. I'm also going to send the obs and forecast
values, instead of obs - forecast, since some of the vector
visualization will actually show both vectors, not just the difference.
It might be a good idea to use a similar format for scalar values, since
it gives us more flexibility; we'll have to see how large these
responses are.