Finish the happy path for diag.wind
I've tested the happy path for the diag.wind function to ensure that it
requests the correct diagnostic data and uses the correct dimensions of
the Dataset to properly construct the return value. There's a lot of
mocking going on in here, and it makes me worried that these tests are
fragile, but it does fully isolate this test from its collaborators.