Implement a function for opening diag files
I've added a function that will take the variable and minimization loop
as parameters and return a Dataset read from the corresponding NetCDF
file. I'll need to refactor get_diagnostic into something that is
specific to fetching temperature and makes use of this function.

I have two placeholder tests for handling missing or bad input files for
open_diagnostic.