Implement VectorVariable.from_vectors
Calculate direction and magnitude of a vector from the (u, v) vectors we
have in our diag files and create a new VectorVariable object from them.
I've checked my test cases against from the NCL documentation:

http://ncl.ucar.edu/Document/Functions/Contributed/wind_direction.shtml

The only difference is that we end up reporting a direction of 90° for a
vector of (0, 0) instead of a direction of 0°. This may end up
mattering, but I think it will be ok, since we'll be internally
consistent when we calculate the Obs - Fcst.