Add a function for generating coordinate pairs
Refactor the list comprehension from wind() into a utility function that
takes an array of longitudes and an array of latitudes and returns a
list of Coordinate tuples. This ensures that wind() can just
orchestrate calls to collaborators and includes no "business logic" for
cleaner testing. It's going to mean a more complicated test with mocks
for wind(), but it also makes it easier to isolate tests for the
coordinates.