Rewrite wind() to use coordinate_pairs_from_vectors
Use the new coordinate_pairs_from_vectors in the wind() function to
generate the list off observation coordinates for the variables.

Rewrote test_wind() to ensure that we create the test data before we
mock the various collaborators. What we were doing originally was
interfering with call counts to the mocked function. Mocking VectorDiag
and then using diag.VectorDiag to create the expected result counted as
a call to the mocked class (of course), so when we try to test the
mocked VectorDiag and assert that it's only called once by the wind()
function, the test fails, because the mock was called twice: once in the
test, once in the wind() function.

It's possible we don't even need to mock the return values here. We
could probably just assert that the return value of one function
(whatever that is) gets passed as an argument to the next collaborator.