Remove all of the complicated mock return values
I was right, we can just test that the calls to one mocked function are
made with the return value of another mocked function, there's no need
for us to set the return value for most of these mocked functions.

I confirmed that

    mock_VectorVariable.return_value != mock_VectorVariable.from_vectors.return_value

so we can be sure that we're testing the behavior of the function with
fine enough detail to catch changes or mistakes like not using the
from_vectors factory.