Remove SvelteKit from deps and configs
Removed Svelte and SvelteKit from our npm scripts and dependencies, as
well as many of the config files that we had for Svelte.

I've also removed Playwright, which we may bring back later for E2E
tests, but for starters I want to keep the list of dependencies as low
as possible, so we'll start with vitest for testing.