Update eslint and bump ecmaVersion to 2022
Private class members are pretty well supported in evergreen browsers
now, so there's no reason not to use them. The version of eslint we had
considered static members and private members on classes to be an error,
so I had to upgrade it and bump the version of ECMAScript it was
expecting.