Expect a loop and value type in the URL for wind diag
Returning raw observations results in a very large response size. It's
made worse when you have 4× observations because you have two
minimization loops and two value types (observation and forecast). So
we're going to switch to requesting each of these separately to reduce
the payload size. We may also need to separate out the coordinates and
look those up via station ID.

For now, only wind operates this way, but I expect we'll want
temperature and all variables to operate this way. This should make it
possible for us to set data sources individually on web components and
make our interface a little more declarative (like adding <img /> tags
is). With HTTP/2, this might actually be an efficient way to load data,
since HTTP/2 can multiplex requests and we can render the pieces of the
data as they come in.