Render wind observations on canvas
I reimplemented the map of wind observations and forecasts using
<canvas>, which is much faster. I also updated the UI to expect
different endpoints for observations and forecasts.

To facilitate fetching data from multiple endpoints, I created a
data-source element that fetches the data and converts it to JSON. It
fires an event (data-changed) that the charts can listen for to
re-render themselves. This declarative API for fetching data was
inspired by the <picture><source></picture> API.

Load times for the map are still pretty long, but the move to <canvas>
ensures that scrolling the app once the data is rendered is fast.