Try to make the map a little more legible
Draw the borders over the data so that they're not obscured by the data.
Scale the size of the dots by the magnitude of the difference between
observed and forecast values. This will call attention to the outliers,
which are the important ones.

I shrank the size of the lines indicating wind direction, but they still
obscure the data because of the density of the observations. This still
needs some work.