Convert the histogram to a web component
I think the charts are good candidates for web components with shadow
DOM to contain the chart. This way they behave more like a replaced
element such as an <img> where you set the src data and get an image
back.

I also decided to change the DiagScalar component to use light DOM and
manage both minimization loops so that it's easier to sync their axes.