Add mouse events for drawing selection
I haven't worked out the right coordinates for the rectangle as it's
being drawn, but I've done the anonymous callback function dance to
ensure that when the methods are called, `this` is our web component and
not the event target and that we can remove the event listeners as
needed.

The coordinate issues could be related to the fact that the canvas is
inside of a shadow root, but I'm not sure yet. I could just be
forgetting how to do this.