Add a VectorObservation dataclass with to_geojson()
To make it easier to work with projections in the UI, I'm going to try
returning our data as GeoJSON. It's going to increase the size of the
payloads significantly, but since we are working with GIS data, this is
a sensible format. We can tackle performance issues as they arise.

This may need to turn into a proper class hierarchy that we can subclass
for vector and scalar variables and serialize to GeoJSON so that all of
our endpoints return data this way.