Rewrite diag.wind to return VectorObservation objects
Updated the tests and rewrote the wind function to return
VectorObservation objects. I'd like to get these calculations refactored
into a pipeline that stores the data in the format we need so that the
API can just fetch the data and format the response, rather than
spending cycles calculating polar coordinates from vectors.