Project the wind map with an Albers projection
I replaced the ridiculous linear scale with an Albers projection. I've
dropped the direction lines from the map for now because they weren't
that useful and they're harder to do with this projection. If they were
more valuable, I'd use an equirectangular projection, but I think they
need to be thought through a little more.