Refactor VectorMapDiag to ChartMap
The bones of the diagnostic view are in place. We display a map next to
a histogram in a single row with one row for each of the two
minimization loops, making it possible to see the entire picture for a
single variable.

Composite components — like the vector diagnostic display which includes
a map and a histogram — are going to be handled by Preact from now on.
We'll use web components to encapsulate components that should be opaque
to the rest of the system, like maps and histograms. So the old
VectorMapdDiag now just handles rendering the map, and the layout of the
map alongside the histograms is handled by the Preact component,
VariableDiagnostic.