Reimplement brushing on the map to filter the histograms
I added a `chart-brush` event to the ChartMap component that fires on
mouseup with the current selection in (lng, lat) pairs. I fire it on
mouseup instead of mousemove right now because it is more performant
this way.

I've converted the VariableDiagnostic component back to a function,
instead of a Component subclass so that we can use hooks to manage the
brushed scalar data. So far, this appears to work rather well, except
that it's not performant enough to allow us to update the histograms
while dragging the selection on the map. That's something we can look at
improving later, perhaps as I become more comfortable with hooks.