Sync selection for both maps
The histograms shared a selection, but the maps didn't, so you could end
up with two different selections on each map and no idea which was
displayed by the histograms (hint: it was whichever you drew last). Now
the selections get updated on both maps regardless of which one you drew
on.