Return GeoJSON from /diag/temperature/
Return a FeatureCollection from /diag/temperature/. I added an
Observation dataclass that can handle both vector and scalar variables,
which is the first step in having a more general way to handle variable
diag files.

I've modified the tests to be integration tests, rather than mocking the
collaborators, which I think makes it a little easier to test. I created
a fixture that takes a filename and an xarray.Dataset and writes a
NetCDF4 file to disk in the data directory during tests (at a tmp path
provided by pytest). I've deleted a number of the unit tests in
test_diag.py because the diag.temperature file is being exercised by the
integration test in test_unified_graphics. We'll see how this shakes
out. It could result in some really complicated test set ups and we end
up going back to mocking, but for the moment it seems like we can run
fewer tests and spend less time mocking and achieve the same results.