Add a menu for selecting the variable
Right now the menu does nothing but display the available variables.
This was just a test to ensure that we can load the data an populate the
select menu with the variables returned by the endpoint. I'll need to do
a little refactoring to make it easier to display the different
variables because we have to distinguish between vector and scalar
variables.