Add a variable type to the Observation class
We'll make use of this on the frontend to decide what display to use for
the variable diagnostics. This will be easier than inspecting the
results on the frontend and trying to make a determination how to handle
things. We have to distinguish scalars and vectors on the backend in
order to process the data correctly, so we might as well just include
that information in our responses.