Rename VariableDiagnostic to DiagnosticView
This component is becoming more complex and I'm beginning to conceive of
it as one view in the application — something that might be one of
several tabs at the top-level of the application. So I'm leaning toward
a naming convention where *View components represent these larger
sections of the app.