Implement basic vector view
Copied over the code for the ScalarVariable component into
VectorVariable and updated it to access the properties on the vector
variable. For now it's just using a histogram for the distribution,
because we have another issue open to implement the contour density to
show distributions of 2D vectors.

There are some errors in the console about getting NaN in some of the
SVG attributes when using the histogram with a vector variable. Not sure
what that's about, but it hopefully won't matter once I swap this out
for a contour density.