Fix a subtle bug in Obs direction calculation
I had reversed the parameters for the vector_direction call, passing in
(v, u) instead of (u, v). Consequently, the values calculated by the
diagnostic function were coming back apparently reversed from what I
expected in my tests.

The wind endpoint test has been updated with values based on the values
used in the unit test for vector_direction. It's very hard to reason
about the inputs and outputs here, it might be wise to have a factory
that takes the vectors for the observation and the forecast and
generates the Dataset with the appropriate "Obs_Minus_Forecast_adjusted"
variables, rather than trying to do that math in your head to calculate
the OMF and writing that directly in the test setup.