Reapply the grid layout from 09242d1
Cherry pick 09242d1b878a46fc4b69745dd2df0393c196d950 to restore that
state. This grid technique results in slightly fewer divs, which is
nice. I couldn't get the responsive vertical layout that I wanted using
the flex technique, so we might as well go back to the grid with fewer
DOM elements.

In order to resize the rows of the layout to fit the screen when it's
large enough, we may have to use media queries (or container queries
when they land). Or we may think about building a grid layout that can
be manually resized so that users can adjust the charts to their liking.

For now, at least, the vertical scrolling is not too bad.