Apply individual grid layouts to the chart rows
I'm trying to work out a way to build a layout that shrinks the rows to
fit the screen when there's enough space, but I'm having trouble making
it work. I thought I could use a flex layout on the outside and just use
the grid to lay out the charts horizontally, but that's not really
paying off.