Refactor common chart features to a base class
There was going to be a lot of copy-pasta from the histogram to the
component for the density plot. Makes more sense to go ahead and
refactor these things now.

So I created a Chart module that contains all of the chart components
and a ChartElement base class that handles the basic markup, styles,
and attributes needed for a resizable chart. The histogram component is
a lot more readable now that it's just a handful of properties specific
to that component plus the render function. I may discover a few more of
these properties that make sense to pull up into ChartElement as I work
on the density plot, too.