Move a few more things into ChartElement
We can override props in subclasses if we need to customize behaviors
when some properties are set, so we can move the data property to the
base class, and the histogram can override the getter and setter to
calculate the mean and standard deviation.

Unfortunately, we have to override the getter, not just the setter, even
though we don't make any changes, because that's just how JS works. If
we only override the setter, then the getter for the subclass returns
undefined for some reason.

I also moved the range property to the ChartElement because that will be
useful for lots of charts. I'm on the fence about a domain property. The
histogram uses a threshold property rather than domain to sync x-axes
between multiple instances, so if the only other chart we end up with is
a density plot, we don't really need the domain on the base class. If we
end up with more plots than that, we should probably have it in the base
class and the histogram can just ignore it.