Batch updates with requestAnimationFrame
I was running into some weird bugs with Preact and the web components
that seems to be related to the render() method in a component growing
too large (i.e. taking too long to run). So instead of just running
render each time something changes, we schedule the render for the next
frame with requestAnimationFrame. If a render is already pending, we
don't have to do anything.

This fixes the weird behavior resulting in SVG with viewboxes that are
`0 0 0 0` (presumably some kind of race condition).