Implement a heatmap for 2D histograms
Experimenting with this instead of the contour density plot. This has
the advantage of showing outliers a little more clearly (although not
much, unless I go for a more intense color scheme).

I think we'll probably stick with this style of chart because it's a
little easier to read. It should also be easier to plot outliers by
replacing the rects where the there's only one observation with a
circle.

Still to do:

- [ ] 1D histograms on each axis
- [ ] Threshold the bin colors
- [ ] Legend for color scheme