Switch to Cividis color scheme for better contrast
The lower end of the scale in D3's interpolateBlues was a little too low
contrast and it wasn't easy to tell where the outliers were. I think
Cividis does a good job of standing out against the background while
also highlighting the denser clusters. (I think that the purple in
Viridis, for example, ends up emphasizing the lower density cells, not
the higher density ones.)