Add a clean script
Before building in dev or production, run clean to delete the build
directory. This way it's easier in development to tell which of the
hashed files is the one you care about.

I also aliased `npm start` to `npm run dev` because I'm tired of typing
the wrong thing. I want a horn here, here, here, and here; you can never
find a horn when you're angry.