Configure coverage tmp directory
c8 puts its report and temporary files in src/coverage by default, which
is annoying. I couldn't figure out how to set the directory using config
options, but there is an environment variable it respects, so I set it
there when we run the tests. It writes to .coverage/ since that's
already in our .gitignore and I don't think we need to see the files
anyway, just the report that's written out to stdout.