Refactor styles into cascade layers
This may not be strictly necessary, but it feels like a best practice to
enclose the framework styles for USWDS in a cascade layer so that I
don't get into specificity battles if I need to override something. So
I'm creating layers for reset, defaults, framework, and layout. Since
utilities are part of the framework, they're actually included there,
although I may look into separating them out later.

I'm having some issue with the layouts, though, because they want to
import USWDS for accessing its tokens and I'm getting some errors from
Sass about it trying to reference `%usa-prose` and not finding it. Not
sure why that's happening. It might be best to just look into generating
custom props for all of the tokens and having my styles rely on those.

I had to add an exception to the stylelint rules to get it to quit
complaining about `@layer`. Maybe that needs an upgrade, since @layer is
well-supported in evergreen browsers now.