Swap units(2) for --space-2 in the layouts
I had to work around the problem of importing the USWDS core for the
tokens in the layout module by creating a custom prop for `units(2)`
that I could use for gap/margin.

Right now it's just hard-coded, but this might actually be a good way to
go about working with the tokens, so I should maybe find a way to
iterate over palettes and auto-generate custom props.