Generate space custom props from palettes
I've figured out how to generate custom properties for USWDS tokens
using the palettes they use for generating utilities. It's a little
risky because it relies on an undocument internal API, so this could end
up breaking in the future, but it's the kind of thing I think we ought
to be able to do.