Apply usa-select styles to our select elements
Add the class to all of our select elements and forward the styles from
USWDS. I added a layer immediately after the framework layer for
overrides so that I can clear out things like the width and margin
settings that USWDS insists on applying to elements regardless of
context. I'm opting for breaking overrides out into separate module
files so that it's easier to see at a glance what components we're
overriding.

I had to put the utility import back into the uswds module because I was
getting another unresolved selector in an @extend rule. I don't know why
this happens when I have multiple forwards, and only for certain
modules. For now, it's probably fine, but there's a chance that I'll
want to set a utility class on a component that has an override, and
then it will be a problem, since the override layer will win against
selector from the component layer.