Add static copy for USWDS image files
As part of the build, we're just copying over the entire img directory
from USWDS. This way we don't have to worry about missing icons or
anything as we add USWDS components and we don't bloat up the repo with
extra files.

We can probably remove the font files from public/ now, if we wanted to,
and just use this to copy them over.