Fix some a11y warnings from Svelte
Tell Svelte to ignore the redundant role on lists. This is here because
we are stripping list styles from it, and Safari won't announce it as a
list unless we add the role back in explicitly.

I've also removed the placeholder anchor tags from the navigation, since
they don't go anywhere and Svelte was mad about the `#` href.