Replace Scalar/VectorVariable with LoopDisplay
Combining the two type components into a single component helps in two
ways:

  1. This should result in more efficient updates, so that we don't
     destroy and recreate the map whenever we change variable types
  2. We don't duplicate layout in our code (things like the heading)

The LoopDisplay component makes the decision of whether to display the
1D or 2D histogram based on the variable type it receives, which keeps
the code in the DiagnosticView component a little easier to read; it
just has to pass down the variable type and let the display do with that
as it will.

There's still a bug in the map that's causing it not to display. I think
I need to distinguish the observations for the map from the values for
the distribution. This will be important later when we introduce
brushing on the histograms, because we don't want to filter the data for
the chart that has a brush applied.