Refactor the layout for the application
Based on a CodePen that creates the responsive layout I want for the
application that has a minimum size for the rows and allows them to
expand to fill available space, and wraps the columns on narrow
viewports.

https://codepen.io/darthmall/pen/yLjqLKq/389a32bcec1d37a46887b433cdb0f379

The codepen works the way I'd expect, but the layout is not working
correctly in the app. The charts are generally taller than I think they
ought to be. This may be the result of the resize observers not firing,
so I'm going to fix that bug and see if things start behaving.