Fix the resize bug
First, I set `contain: strict` on both the host element of the web
component and on the container with the resize observer. This is
significantly more elegant than setting `position: relative; overflow:
hidden`, etc.

Second, I set up the wrapper on the ChartMap with the fixed resize code,
because I think that's what was causing the rows to grow oddly. Since
one of the charts in the row was misbehaving, the whole row misbehaved.

Using a div container for the resize observer and setting `contain:
strict` on everything causes the rows to grow and shrink as needed to
fill the available space, but no smaller than the row's min height.